refactor(mediawiki): extract client.go auth/cache/helpers into sibling files (deslop #59)
